pyproject.toml
[tool.poetry]
name = "ai-memory-service"
version = "0.1.0"
description = "Microserviço intermediário entre API Gateway e Flowise com memória robusta usando Graphiti"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-decouple = "^3.8"
httpx = "^0.25.2"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
sqlalchemy = "^2.0.23"
redis = "^5.0.1"
celery = "^5.3.4"
graphiti-core = "^0.3.3"
neo4j = "^5.14.1"
structlog = "^23.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
